Add basic tests for bls12_381
